Move type to matching filename
