+++ src/Serilog.Sinks.Console/Sinks/SystemConsole/Output/OutputTemplateRenderer.cs
@@ -19,7 +19,6 @@ using Serilog.Events;
 using Serilog.Formatting;
 using Serilog.Formatting.Display;
 using Serilog.Parsing;
-using Serilog.Sinks.SystemConsole.Rendering;
 using Serilog.Sinks.SystemConsole.Themes;
 
 namespace Serilog.Sinks.SystemConsole.Output
@@ -83,55 +82,6 @@ namespace Serilog.Sinks.SystemConsole.Output
 
             foreach (var renderer in _renderers)
                 renderer.Render(logEvent, output);
-        }
-    }
-
-    class EventPropertyTokenRenderer : OutputTemplateTokenRenderer
-    {
-        readonly ConsoleTheme _theme;
-        readonly PropertyToken _token;
-        readonly IFormatProvider _formatProvider;
-
-        public EventPropertyTokenRenderer(ConsoleTheme theme, PropertyToken token, IFormatProvider formatProvider)
-        {
-            _theme = theme;
-            _token = token;
-            _formatProvider = formatProvider;
-        }
-
-        public override void Render(LogEvent logEvent, TextWriter output)
-        {
-            // If a property is missing, don't render anything (message templates render the raw token here).
-            LogEventPropertyValue propertyValue;
-            if (!logEvent.Properties.TryGetValue(_token.PropertyName, out propertyValue))
-            {
-                Padding.Apply(output, "", _token.Alignment);
-                return;
-            }
-
-            var _ = 0;
-            using (_theme.Apply(output, ConsoleThemeStyle.SecondaryText, ref _))
-            {
-                var writer = _token.Alignment.HasValue ? new StringWriter() : output;
-
-                // If the value is a scalar string, support some additional formats: 'u' for uppercase
-                // and 'w' for lowercase.
-                if (propertyValue is ScalarValue sv && sv.Value is string literalString)
-                {
-                    var cased = Casing.Format(literalString, _token.Format);
-                    writer.Write(cased);
-                }
-                else
-                {
-                    propertyValue.Render(writer, _token.Format, _formatProvider);
-                }
-
-                if (_token.Alignment.HasValue)
-                {
-                    var str = writer.ToString();
-                    Padding.Apply(output, str, _token.Alignment);
-                }
-            }
         }
     }
 }
